Revert "Fix/merge manual estilos"
