+++ app/(mainScreen)/_layout.tsx
@@ -16,10 +16,36 @@ export default function MainScreenLayout() {
 
               
         }}>
-            <Tabs.Screen name="main" options={{ headerShown: false, }} />
-            <Tabs.Screen name="ajuda"  options={{ headerShown: false }} />
-            <Tabs.Screen name="perfil" options={{ headerShown: false }} />
-            {/*<Tabs.Screen name="chat" options={{headerShown: false,  headerStyle: { backgroundColor: '#4CAF50' }}} />*/}
+            <Tabs.Screen name="main" 
+            options={{ 
+              headerShown: false, 
+              tabBarIcon: ({ color, size }) => (
+                <Image source={require("@/assets/icons/HomeIcon.png")} style={{ width: 40, height: 40, tintColor: "white" }} />
+              )  
+            }} />
+            <Tabs.Screen name="ajuda"
+              options={{ 
+                headerShown: false,
+                tabBarIcon: ({ color, size }) => (
+                  <Image source={require("@/assets/icons/AjudaIcon.png")} style={{ width: 40, height: 40, tintColor: "white" }} />
+                )  
+              }}/>
+            <Tabs.Screen name="perfil" 
+              options={{
+                headerShown: false,
+                tabBarIcon: ({ color, size }) => (
+                  <Image source={require("@/assets/icons/PerfilIcon.png")} style={{ width: 30, height: 33, tintColor: "white" }} />
+                ) 
+
+               }} />
+            <Tabs.Screen name="chat" 
+              options={{
+                headerShown: false,  
+                headerStyle: { backgroundColor: '#4CAF50' },
+                tabBarIcon: ({ color, size }) => (
+                  <Image source={require("@/assets/icons/SoeSopIcon.png")} style={{ width: 40, height: 40, tintColor: "white" }} />
+                ) 
+              }} />
           </Tabs>
       </GestureHandlerRootView>
     </AuthProvider>
