+++ app/(mainScreen)/chats/[otherUserId].jsx
