+++ styles/frequenciaStyles.js
@@ -1,6 +1,4 @@
-import { StyleSheet, Dimensions } from "react-native";
-
-const screenWidth = Dimensions.get('window').width;
+import { StyleSheet } from "react-native";
 
 const frequenciaStyles = StyleSheet.create({
 
@@ -11,25 +9,28 @@ const frequenciaStyles = StyleSheet.create({
         flex:1,
         widht:"100%",
         height:"100%",
-        marginTop:50,
+        marginTop:100,
     
     },
 
     title:{
         fontSize:25,
+        fontFamily:"Arial",
         fontWeight:"bold",
         color:"green"
     },
 
     line:{
-        width:screenWidth,
+        width:480,
         height:1,
         marginBottom:20,
         marginVertical:30,
         backgroundColor: "gray",
+        marginLeft:40
     },
 
     infos:{
+        fontFamily:"Arial",
         fontWeight:"bold",
         fontSize:25,
         marginTop:20,
